+++ src/Elems/AsideElems/StrategyControl.tsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import { _expect } from '../../utils';
 import { _ReactProps } from '../../Types';
 import Control from '../Control';
 
@@ -12,18 +11,8 @@ type StrategyControlProps = Readonly<{
  * Strategy control
  * When a user clicks on the control... something happens,
  * like a strategy being run against the sudoku.
- *
- * @requiredProps
- * - onClick
- * - name
  */
 export default class StrategyControl extends React.Component<StrategyControlProps> {
-   constructor(props: StrategyControlProps) {
-      _expect(StrategyControl, props).toHaveProperties("onClick", "name")
-
-      super(props)
-   }
-
    render() {
       return (
          <Control