+++ src/Elems/AsideElems/StrategyLabel.tsx
@@ -1,7 +1,6 @@
 // @flow
 
 import React from 'react';
-import { _expect } from '../../utils';
 import ExternalLink from '../ExternalLink';
 
 export type StrategyLabelProps = Readonly<{
@@ -16,22 +15,16 @@ export type StrategyLabelProps = Readonly<{
  * Really it's just the text inside the StrategyItem,
  * besides StrategyResult
  */
-export default class StrategyLabel extends React.Component<StrategyLabelProps> {
-   constructor(props: StrategyLabelProps) {
-      _expect(StrategyLabel, props).toHaveProperties("name")
-      super(props)
-   }
-
+export default class StrategyLabel extends React.PureComponent<StrategyLabelProps> {
    render() {
-      let content = <span className="StrategyLabel" id={this.props.id}>{this.props.name}</span>
       if (this.props.href) {
-         content = (
+         return (
             <ExternalLink className="StrategyLabel" id={this.props.id} href={this.props.href}>
                {this.props.name}
             </ExternalLink>
          )
       }
 
-      return content
+      return <span className="StrategyLabel" id={this.props.id}>{this.props.name}</span>
    }
 }