+++ src/Elems/AsideElems/StrategyToggler.tsx
@@ -1,6 +1,5 @@
 
 import React from 'react';
-import { _expect } from '../../utils';
 
 type StrategyTogglerProps = Readonly<{
    callback: React.ChangeEventHandler
@@ -13,15 +12,9 @@ type StrategyTogglerState = Readonly<{
 
 /**
  * Turns a strategy off or on
- *
- * @requiredProps
- * - callback
- * - id
  */
 export default class StrategyToggler extends React.Component<StrategyTogglerProps, StrategyTogglerState> {
    constructor(props: StrategyTogglerProps) {
-      _expect(StrategyToggler, props).toHaveProperties("callback", "id")
-
       super(props)
 
       this.state = {