+++ src/Elems/Main.tsx
@@ -4,24 +4,18 @@ import React from 'react'
 import Sudoku, { BaseSudokuProps } from './MainElems/Sudoku'
 import DataContainer from './MainElems/Data'
 import Coords from './MainElems/Coords'
-import { _expect } from '../utils'
 
 type MainProps = Readonly<{
    propsPassedDown: BaseSudokuProps
 }>
 
 /**
- * The "main" component, which includes the Data, Sudoku, and Coords components
+ * The "main" component, which is just the sudoku parts for now.
  *
- * @requiredProps
- * - whenSudokuConstructs
+ * Currently the parts are Coords, Sudoku, and Data
+ * TODO: Remove Data
  */
 class Main extends React.Component<MainProps> {
-   constructor(props: MainProps) {
-      _expect(Main, props).toHaveProperty("propsPassedDown")
-      super(props)
-   }
-
    render() {
       return (
          <main className="App-main">