+++ src/utils.test.ts
@@ -1,5 +1,5 @@
 import React from "react"
-import { convertArrayToEnglishList, _expect } from "./utils"
+import { convertArrayToEnglishList } from "./utils"
 
 test("convertArrayToEnglishList", () => {
    expect(() => convertArrayToEnglishList([])).toThrow(TypeError)
@@ -8,13 +8,4 @@ test("convertArrayToEnglishList", () => {
    expect(convertArrayToEnglishList([1, 2, 3])).toBe("1, 2, and 3")
    expect(convertArrayToEnglishList([1, 2, 3, 4])).toBe("1, 2, 3, and 4")
    expect(convertArrayToEnglishList("ABCDEFGHIJKLMNOPQRSTUVWXYZ".split(""))).toBe("A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, P, Q, R, S, T, U, V, W, X, Y, and Z")
-})
-
-test("_expect", () => {
-   class SomeComponent extends React.Component {}
-
-   expect(() => _expect(SomeComponent, { apple: 7 }).toHaveProperty("apple")).not.toThrow()
-   expect(() => _expect(SomeComponent, { apple: 7, jump: 5 }).toHaveProperties("apple", "jump")).not.toThrow()
-   expect(() => _expect(SomeComponent, { jump: 5 }).toHaveProperty("apple")).toThrow()
-   expect(() => _expect(SomeComponent, { apple: 7 }).toHaveProperties("apple", "jump")).toThrow()
 })